chore: move obs-mcp to root repo
