--- .ci-operator.yaml
@@ -1,4 +1,0 @@
-build_root_image:
-  name: nodejs-18
-  namespace: openshift
-  tag: latest
