--- .vscode/settings.json
@@ -1,21 +1,0 @@
-{
-  "typescript.tsdk": "./node_modules/typescript/lib",
-  "search.exclude": {
-    "**/node_modules": true
-  },
-  "files.watcherExclude": {
-    "**/node_modules/**": true
-  },
-  "files.associations": {
-    "**/console-extensions.json": "jsonc"
-  },
-  "yaml.schemas": {
-    "https://raw.githubusercontent.com/RedHat-UX/next-gen-ui-agent/refs/heads/main/spec/config/agent_config.schema.json":  ["ngui_*.yaml", "ngui_*.yml"]
-  },
-  "json.schemas": [
-    {
-      "fileMatch": ["**/console-extensions.json"],
-      "url": "./node_modules/@openshift-console/dynamic-plugin-sdk-webpack/schema/console-extensions.json"
-    }
-  ]
-}
