+++ cmd/obs-mcp/main.go
@@ -9,10 +9,10 @@ import (
 
 	"github.com/prometheus/common/promslog"
 
-	"github.com/inecas/obs-mcp/pkg/k8s"
-	"github.com/inecas/obs-mcp/pkg/mcp"
-	"github.com/inecas/obs-mcp/pkg/prometheus"
 	"github.com/mark3labs/mcp-go/server"
+	"github.com/rhobs/obs-mcp/pkg/k8s"
+	"github.com/rhobs/obs-mcp/pkg/mcp"
+	"github.com/rhobs/obs-mcp/pkg/prometheus"
 )
 
 const (
