+++ manifests/03_mcp_service.yaml
