--- obs-mcp/.gitignore
@@ -1,1 +1,0 @@
-.claude
