+++ pkg/mcp/tools.go
