+++ pkg/mcp/tools_test.go
